Graphe polynôme spect vs presse
